feat(profiling): remap sourcemaps using webpack plugin
